Add validation for updating.

+++ update_and_overwrite.sh
@@ -1,9 +1,20 @@
 #!/bin/sh
 
-echo -e '\nUpdate repository:'
+echo -e "\nUpdate repository and overwrite local changes [y/n]?"
+read ans
+
+if [ $ans = y -o $ans = Y -o $ans = yes -o $ans = Yes -o $ans = YES ]
+then
 
 echo -e '\ngit fetch origin master'
 git fetch origin master
 
 echo -e '\ngit reset --hard origin/master'
 git reset --hard origin/master'
+
+fi
+
+if [ $ans = n -o $ans = N -o $ans = no -o $ans = No -o $ans = NO ]
+then
+echo "Not updating."
+fi
